Beginning of UI Implementation
So this is the framework code and my current progress on making the program run through a UI.  I started reading up on tkinter before the prof even introduced it, so I'm happy my early work wasn't for nothing!
I'm having a difficult time grasping where the variables are in the program, what variables need to be part of what function... things like that.  Having a UI for the program breaks up the code even more!  You will see near the bottom of the code, there are a bunch of .bind('<button-1>'...) these are on click events for the UI buttons.  'button-1' is a click of the user's left mouse button on that element (ex: the code says for something to happen when the user left-clicks the button labeled "Add weekly payments", etc).
I don't fully understand how to use them; thinking the function should be called upon a registered left click of the button in the ui (hense the term on click event), so I'm working on learning the syntax for these.
PLEASE NOTE:  root.mainloop() MUST BE THE VERY LAST LINE OF THE CODE.  THIS CALLS FOR THE UI TO OPEN (in jupyter notebook, the ui is in a separate window from the code).  If that line isn't there, the code will NOT run!